HTML scenario-help feature (compatible). Config changes (read "description")
Add (and document) a feature where a third arg in :HELP-TEXT (mis-documented, too) can specify an HTML file which will be browsed (core and Mac implementation, Windows pending).  Create such help for Myrtle Avenue.  Upgrade app deployment to Mojave (10.14) because filesystem::path will not work earlier than that.  This broke all uses of NSOn/OffState which had to be upgraded to NSControlStateValueOn/Off in a handful of modules.