TLEdit - new Windows Dialog emulation
(all except Text String dialog, which is complex - when that is done, the old stuff will be removed).  New Windows dialog system relies on "identifier" strings in XIB controls, equal to Windows resource_id names.  Controls didn't have them in 2014, when this code was first built, and an error-prone system based upon title string matching and tab order chasing substituted for it.  Soon to be history.