TLEdit bug fix involving exit lights.  See descr.
When a segment was split by the introduction of a new joint, the exit light associated with the end of the segment which is transferred to the new segment was not informed of the ownership change.  This caused the joint to be undeletable and attempts to delete the exit light to crash.