Fix Windows command line parsing.
Use correct Windows-only technology (MultiByteToWide and vice versa) to prepare and receive data for the command-line parsing API, which is Unicode(UTF-16) only.  The entire previous strategy, c++ UTF8/16, was nonsense. Windows has no truck with UTF8. Well, it went deprecated anyway, and the replacement I wrote in pure C++ didn't work for that reason.  This solution is tested, putting an é in an argument made it through to the error message!  What cruft, CP_ACP in this day and age!

By the way, std::min cannot be used on Windows!